% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbst.R
\name{FAST}
\alias{FAST}
\title{Stability calculation of variety}
\usage{
FAST(loading, scores)
}
\arguments{
\item{loading}{The rotated loading matrix.}

\item{scores}{The score matrix with the rownames containing the names of the
varieties.}
}
\value{
A named list containing overall performance (OP) and
root mean square deviation (RSMD)
}
\description{
The function also performs a check taht the
estimated loadings for the first factor is all positive.
}
\seealso{
rotate
}
