% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{loadings}
\alias{loadings}
\title{Loading matrix}
\format{
A matrix with 4 columns (factors) and 8 rows (environment).
}
\usage{
loadings
}
\description{
A loading matrix from CAIGE 2018 data supplied by Manigben Kulai Amadu.
}
\keyword{datasets}
