% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbst-package.R
\docType{package}
\name{pbst-package}
\alias{pbst}
\alias{pbst-package}
\title{pbst: Plant Breeding Selection Tools}
\description{
Implements tools to assist selection from 
 multi-environment trial data of plant breeding experiments
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/emitanaka/pbst}
  \item Report bugs at \url{https://github.com/emitanaka/pbst/issues}
}

}
\author{
\strong{Maintainer}: Emi Tanaka \email{dr.emi.tanaka@gmail.com} (\href{https://orcid.org/0000-0002-1455-259X}{ORCID})

}
\keyword{internal}
