% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scores}
\alias{scores}
\title{Score matrix}
\format{
A matrix with 4 columns (scores) and 315 rows (varieties).
}
\usage{
scores
}
\description{
A score matrix from CAIGE 2018 data supplied by Manigben Kulai Amadu.
}
\keyword{datasets}
